mimxrt: Enabling i2s for MIMXRT10xx_DEV boards.
Uing the WM8960 codec requires a different setting for read().
For the moment it is done using a compile switch. But a init() option
would be better.
Using the WM8960 for recording requires the wm8960 chip to run in
master mode. That is a little bit inconveient, since sample rate
and bit size has to be specified twice.